arreglo error de fk constraint incorrectly formed

+++ app/models/category.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, SMALLINT
 from app.db import db
 from sqlalchemy.orm import validates, relationship
 from sqlalchemy.sql.expression import desc, select
@@ -7,7 +7,7 @@ class Category(db.Model):
     """Define una entidad de tipo Category que se corresponde con el table categories"""
 
     __tablename__ = "categories"
-    id = Column(Integer, primary_key=True)
+    id = Column(SMALLINT, primary_key=True)
     name = Column(String(30), unique=True)
 
     def __init__(self, name=None):
